cosmwasm/enclaves/execute almost compiles
