+++ cosmwasm/enclaves/execute/Cargo.toml
@@ -14,7 +14,6 @@ default = ["SGX_MODE_SW"]
 SGX_MODE_SW = []
 SGX_MODE_HW = []
 production = ["SGX_MODE_HW", "log/max_level_warn", "log/release_max_level_warn"]
-build-protobuf = ["protoc-rust", "dirs"]
 debug-print = []
 test = []
 
@@ -22,36 +21,34 @@ test = []
 # We always compile to the "sgx" target, so this will always be false.
 # when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
 [target.'cfg(not(target_env = "sgx"))'.dependencies]
-sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_tstd = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = [
     "backtrace"
 ] }
 
 [dependencies]
+sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_tse = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_rand = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-sgx_trts = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-sgx_tseal = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_tcrypto = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-aes-siv = "=0.6.2"
+
 enclave-ffi-types = { path = "../ffi-types" }
 
 enclave_contract_engine = { path = "../shared/contract-engine", features = ["query-only"] }
+enclave_crypto = { path = "../shared/crypto" }
 enclave_utils = { path = "../shared/utils" }
+
+serde = { git = "https://github.com/mesalock-linux/serde-sgx", features = [
+  "derive"
+] }
+serde_json = { git = "https://github.com/mesalock-linux/serde-json-sgx" }
+ctor = "0.1.13"
+derive_more = "0.99"
 
 pwasm-utils = { version = "0.12.0", default-features = false }
 parity-wasm = { version = "0.41.0", default-features = false }
-serde = { git = "https://github.com/mesalock-linux/serde-sgx", features = [
-    "derive"
-] }
-serde_json = { git = "https://github.com/mesalock-linux/serde-json-sgx" }
+
 base64 = { rev = "dc7389e10817b078f289386b3b6a852ab6c4c021", git = "https://github.com/mesalock-linux/rust-base64-sgx" }
-ctor = "0.1.13"
-bech32 = "0.7.2"
-derive_more = "0.99"
-sha2 = "0.8.1"
-ring = { git = "https://github.com/mesalock-linux/ring-sgx", tag = "v0.16.5" }
-secp256k1 = "0.19.0"
+
 # for attestation
 chrono = { git = "https://github.com/mesalock-linux/chrono-sgx" }
 num-bigint = { git = "https://github.com/mesalock-linux/num-bigint-sgx" }
@@ -60,31 +57,8 @@ httparse = { version = "1.3", default-features = false }
 itertools = { version = "0.8", default-features = false, features = [] }
 bit-vec = { version = "0.6", default-features = false }
 lazy_static = "1.4"
-x25519-dalek = { version = "=1.2.0", default-features = false, features = [
-    "u64_backend"
-] }
 hex = "0.4.2"
-ripemd160 = "0.9.1"
 log = "0.4.8"
-lru = { version = "0.6", default-features = false }
-protobuf = "2.22.0"
-
-[dependencies.prost]
-git = "https://github.com/mesalock-linux/prost-sgx"
-rev = "cd3103a6d45cf7a43b6c1c5e4223428097d1c547"
-default-features = false
-features = ["prost-derive"]
-
-[dependencies.wasmi]
-git = "https://github.com/paritytech/wasmi"
-rev = "84d2764594d80425373bf4949a58fa3df3d624c3"
-default-features = false
-features = ["core"]
-
-[dependencies.base58]
-git = "https://github.com/mesalock-linux/rust-base58-sgx"
-package = "rust-base58"
-rev = "13fb3e0a543690e6e19332f37ba85fd74c56cb2f"
 
 [dependencies.webpki]
 git = "https://github.com/mesalock-linux/webpki"
@@ -108,7 +82,3 @@ features = ["dangerous_configuration", "mesalock_sgx"]
 
 [dev-dependencies]
 sgx_edl = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
-
-[build-dependencies]
-protoc-rust = { version = "=2.25", optional = true }
-dirs = { version = "3.0", optional = true }
