+++ cosmwasm/enclaves/execute/Makefile
@@ -93,7 +93,7 @@ RustEnclave_Link_Flags := -Wl,--no-undefined -nostdlib -nodefaultlibs -nostartfi
 
 
 Rust_Crate_Source := $(shell find -type f | grep -vP '(\.so|\.o|\.a)$$' | grep -vP '^\./target' | grep -v Enclave_t)
-Rust_Target_Path := $(CURDIR)/xargo
+Rust_Target_Path := $(CURDIR)/../xargo
 
 ifeq ($(SGX_MODE), SW)
 	Rust_Flags := "-Z force-unstable-if-unmarked -C target-feature=+aes,+ssse3"
