+++ cosmwasm/enclaves/execute/src/lib.rs
@@ -1,38 +1,15 @@
 // Trick to get the IDE to use sgx_tstd even when it doesn't know we're targeting SGX
 #[cfg(not(target_env = "sgx"))]
 extern crate sgx_tstd as std;
-// This annotation is here to trick the IDE into ignoring the extern crate, and instead pull in sgx_types from our
-// Cargo.toml. By importing sgx_types using `extern crate` but without letting it resolve in Cargo.toml when compiling
-// to SGX, we make the compiler pull it in from the target root, which contains the sgx_types listed in Xargo.toml.
-// This in turn silences errors about using the same types from two versions of the same crate.
-// (go ahead, try to remove this block and change the Cargo.toml import to a normal one)
-#[cfg(target_env = "sgx")]
-extern crate sgx_types;
 
 #[allow(unused_imports)]
 use ctor::*;
 use log::LevelFilter;
 
-use crate::logger::*;
+use enclave_utils::logger::{SimpleLogger, LOG_LEVEL_ENV_VAR};
 
-mod macros;
-
-pub mod exports;
-pub mod imports;
-pub mod logger;
-mod oom_handler;
-mod recursion_depth;
 pub mod registration;
 use std::env;
-
-mod consts;
-mod cosmwasm;
-mod crypto;
-mod proto;
-mod results;
-mod storage;
-mod utils;
-mod wasm;
 
 mod tests;
 
@@ -76,7 +53,7 @@ fn set_log_level_or_default(default: LevelFilter, max_level: LevelFilter) {
     let mut log_level = default;
 
     if let Some(env_log_level) =
-        log_level_from_str(&env::var(consts::LOG_LEVEL_ENV_VAR).unwrap_or_default())
+        log_level_from_str(&env::var(LOG_LEVEL_ENV_VAR).unwrap_or_default())
     {
         // We want to make sure log level is not higher than WARN in production to prevent accidental secret leakage
         if env_log_level <= max_level {
