+++ cosmwasm/enclaves/execute/src/registration/seed_exchange.rs
@@ -1,8 +1,9 @@
 use log::*;
+
 use sgx_types::{sgx_status_t, SgxResult};
 
-use crate::consts::ENCRYPTED_SEED_SIZE;
-use crate::crypto::{
+use enclave_crypto::consts::ENCRYPTED_SEED_SIZE;
+use enclave_crypto::{
     AESKey, Keychain, SIVEncryptable, Seed, KEY_MANAGER, PUBLIC_KEY_SIZE, SEED_KEY_SIZE,
 };
