+++ cosmwasm/enclaves/shared/crypto/src/consts.rs
@@ -73,7 +73,6 @@ pub const CONSENSUS_IO_EXCHANGE_KEYPAIR_DERIVE_ORDER: u32 = 2;
 pub const CONSENSUS_STATE_IKM_DERIVE_ORDER: u32 = 3;
 pub const CONSENSUS_CALLBACK_SECRET_DERIVE_ORDER: u32 = 4;
 
-pub const LOG_LEVEL_ENV_VAR: &str = "LOG_LEVEL";
 pub const SCRT_SGX_STORAGE_ENV_VAR: &str = "SCRT_SGX_STORAGE";
 
 const DEFAULT_SGX_SECRET_PATH: &str = "/opt/secret/.sgx_secrets/";
